services/voice/pyproject.toml
[project]
name = "moc-voice"
version = "0.1.0"
description = "Voice-first pipeline: SileroVAD → GroqSTT → Claude CLI → CartesiaTTS via Pipecat + Daily.co"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pipecat-ai[daily,groq,silero,cartesia]>=0.0.102",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.27.0",
]
